# Add better guards again re-entrance in ByteToMessageDecoder