@normanmaurer 2018-04-28
This needs more tests and we may want to split it in two PRs (while the second needs the first)...

That said here are some numbers for allocations:


Before: `1000_reqs_1_conn.total_allocations: 61113`
After: `1000_reqs_1_conn.total_allocations: 58116`

And benchmark via wrk:

Before:
```
Running 10s test @ http://localhost:8888
  4 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   104.15ms  206.11ms   1.99s    91.22%
    Req/Sec   132.16k    26.40k  324.24k    89.22%
  5258220 requests in 10.03s, 250.73MB read
  Socket errors: connect 0, read 154, write 0, timeout 22
Requests/sec: 524473.37
Transfer/sec:     25.01MB
```

After:
```
Running 10s test @ http://localhost:8888
  4 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   103.21ms  204.95ms   1.98s    88.99%
    Req/Sec   135.99k    25.68k  304.99k    88.75%
  5425784 requests in 10.03s, 258.72MB read
  Socket errors: connect 0, read 157, write 0, timeout 12
Requests/sec: 540810.81
Transfer/sec:     25.79MB
```