@normanmaurer 2018-04-28
Also tests pass in docker locally... we may need to update timeouts