@normanmaurer 2018-05-04
Let me update the PR to only include the `ByteToMessageDecoder` changes.