@normanmaurer 2018-05-04
@weissi @Lukasa PTAL again... Will add some more tests as well.