@normanmaurer 2018-05-04
@Lukasa haha you were faster then me (and I did not refresh yet). Yep will add a few tests.