@weissi 2018-05-04
@normanmaurer awesome!

```
18:30:29 info: 1000_reqs_1_conn: total number of mallocs: 46730
18:30:29 info: 1_reqs_1000_conn: allocations not freed: 0
18:30:29 info: 1_reqs_1000_conn: total number of mallocs: 775997
18:30:29 info: ping_pong_1000_reqs_1_conn: allocations not freed: 0
18:30:29 info: ping_pong_1000_reqs_1_conn: total number of mallocs: 4513
18:30:29 info: bytebuffer_lots_of_rw: allocations not freed: 0
18:30:29 info: bytebuffer_lots_of_rw: total number of mallocs: 3011
18:30:29 info: future_lots_of_callbacks: allocations not freed: 0
18:30:29 info: future_lots_of_callbacks: total number of mallocs: 99001
```

Please edit the docker compose files and lower the limits once again.